jcsmesquita/git-money/requirements.txt
Flask==0.10.1
Flask-Hookserver==1.0.0
requests==2.2.1
Pillow==3.1.0
commonregex
tweepy
